Handled min column width.
